feat: 수정 완료 #9 - controller 명칭 수정 - Repository Integer -> Long 수정 - Entity 컬럼 삭제

+++ backend/src/main/java/com/shallwecode/backend/problem/domain/repository/ProblemRepository.java
@@ -3,6 +3,6 @@ package com.shallwecode.backend.problem.domain.repository;
 import com.shallwecode.backend.problem.domain.aggregate.Problem;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProblemRepository extends JpaRepository<Problem, Integer> {
+public interface ProblemRepository extends JpaRepository<Problem, Long> {
 
 }
