+++ backend/src/main/java/com/shallwecode/backend/problem/domain/repository/TestcaseRepository.java
@@ -3,6 +3,6 @@ package com.shallwecode.backend.problem.domain.repository;
 import com.shallwecode.backend.problem.domain.aggregate.Testcase;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TestcaseRepository extends JpaRepository<Testcase, Integer> {
+public interface TestcaseRepository extends JpaRepository<Testcase, Long> {
 
 }
